Move character endpoints into their own module
